feat(instructions): Separation by usage
The `Instruction` subclasses were divided into 3 category:
1. `Preparation`;
2. `Gate`;
3. `Measurement`.

This is justified by the fact that e.g. we want to force the users to
specify the instructions in the above order and write validations. For
the validations, the existence of these subclasses is convenient.
