+++ piquasso/api/instruction.py
@@ -20,6 +20,7 @@ import numpy as np
 
 from .mode import Q
 from piquasso.core import _mixins, _registry
+from piquasso.api.errors import InvalidProgram
 
 if typing.TYPE_CHECKING:
     from piquasso.api.program import Program
@@ -127,3 +128,24 @@ class Instruction(_mixins.DictMixin, _mixins.RegisterMixin, _mixins.CodeMixin):
             and
             self.params == other.params
         )
+
+
+class Preparation(Instruction):
+    """Base class for preparations."""
+
+
+class Gate(Instruction):
+    """Base class for gates."""
+
+
+class Measurement(Instruction):
+    r"""Base class for all measurements."""
+
+    def _apply_to_program_on_register(self, program: "Program", register: Q) -> None:
+        if any(
+            isinstance(instruction, type(self))
+            for instruction in program.instructions
+        ):
+            raise InvalidProgram("Measurement already registered.")
+
+        super()._apply_to_program_on_register(program, register)
