+++ piquasso/instructions/measurements.py
@@ -25,7 +25,7 @@
 import numpy as np
 
 from piquasso.api.errors import InvalidParameter
-from piquasso.api.measurement import Measurement
+from piquasso.api.instruction import Measurement
 from piquasso._math.linalg import is_positive_semidefinite
 from piquasso._math.symplectic import symplectic_form
